ок
ок

+++ storage/models.py
@@ -6,7 +6,7 @@ from orders.models import Order
 class StorageItemTextile(models.Model):
     item = models.ForeignKey(Textile, null=True, on_delete=models.CASCADE, verbose_name='Текстиль', )
     type_p = models.CharField(verbose_name='Тип', max_length=100, blank=True, null=True, default='Текстиль')
-    quantity = models.IntegerField(default=0, verbose_name='Количество')
+    quantity = models.FloatField(default=0, verbose_name='Количество')
     price = models.FloatField(verbose_name='Закупочная стоимость', default=0)
     date_created = models.DateTimeField(verbose_name='Дата создания', auto_now_add=True, null=True)
 
@@ -16,7 +16,7 @@ class StorageItemTextile(models.Model):
 class StorageItemTextileReserve(models.Model):
     order = models.ForeignKey(Order, null=True, on_delete=models.CASCADE, verbose_name='Заказ')
     item = models.ForeignKey(StorageItemTextile, null=True, on_delete=models.CASCADE, verbose_name='Текстиль')
-    quantity = models.IntegerField(default=0, verbose_name='Количество')
+    quantity = models.FloatField(default=0, verbose_name='Количество')
     date_created = models.DateTimeField(verbose_name='Дата создания', auto_now_add=True, null=True)
 
     def __str__(self):
@@ -26,7 +26,7 @@ class StorageItemTextileReserve(models.Model):
 class StorageItemCornice(models.Model):
     item = models.ForeignKey(Cornice, null=True, on_delete=models.CASCADE, verbose_name='Карнизы', )
     type_p = models.CharField(verbose_name='Тип', max_length=100, blank=True, null=True, default='Карнизы')
-    quantity = models.IntegerField(default=0, verbose_name='Количество')
+    quantity = models.FloatField(default=0, verbose_name='Количество')
     reserve = models.IntegerField(default=0, verbose_name='Резерв')
     price = models.FloatField(verbose_name='Закупочная стоимость', default=0)
     date_created = models.DateTimeField(verbose_name='Дата создания', auto_now_add=True, null=True)
